move example from merge
