+++ src/components/examples/example-component.tsx
